requirements.txt
customtkinter
pygame
pillow
